+++ HISTORY.md
@@ -1,5 +1,13 @@
 # Revision Change History
 
+## [0.6]
+
+### Additions
+- Battery powered devices will now attempt to download the all link
+  database if it doesn't exist locally when a broadcast message from
+  that device is seen (since the device is awake)
+
+
 ## [0.5.1] - 2017-12-07
 
 ### Fixed